#60 Refactor for merge
