+++ src/main/java/com/solvd/qaprotours/service/impl/Scheduler.java
@@ -1,7 +1,7 @@
 package com.solvd.qaprotours.service.impl;
 
 import com.solvd.qaprotours.config.kafka.KafkaMessage;
-import com.solvd.qaprotours.config.kafka.MessageSenderImpl;
+import com.solvd.qaprotours.config.kafka.MessageSender;
 import com.solvd.qaprotours.domain.MailData;
 import com.solvd.qaprotours.domain.MailType;
 import com.solvd.qaprotours.service.TicketService;
@@ -27,7 +27,7 @@ public class Scheduler {
     private final UserClient userClient;
     private final MailDataMapper mailDataMapper;
     private final UserMapper userMapper;
-    private final MessageSenderImpl<MailDataDto> messageSender;
+    private final MessageSender<MailDataDto> messageSender;
 
     /**
      * Find all booked tickets and send notification emails to users.
@@ -122,8 +122,8 @@ public class Scheduler {
                                                 params);
                                         MailDataDto dto =
                                                 mailDataMapper.toDto(mailData);
-                                        KafkaMessage message
-                                                = new KafkaMessage();
+                                        KafkaMessage<MailDataDto> message
+                                                = new KafkaMessage<>();
                                         message.setTopic("mail");
                                         message.setPartition(0);
                                         message.setKey(ticket.getUserId());
