+++ src/test/java/com/solvd/qaprotours/config/TestConfig.java
@@ -1,8 +1,11 @@
 package com.solvd.qaprotours.config;
 
+import com.solvd.qaprotours.config.kafka.MessageSender;
+import com.solvd.qaprotours.config.kafka.MessageSenderImpl;
 import com.solvd.qaprotours.repository.TicketRepository;
 import com.solvd.qaprotours.repository.TourRepository;
 import com.solvd.qaprotours.service.AuthService;
+import com.solvd.qaprotours.service.HotelService;
 import com.solvd.qaprotours.service.JwtService;
 import com.solvd.qaprotours.service.TicketService;
 import com.solvd.qaprotours.service.TourService;
@@ -13,6 +16,7 @@ import com.solvd.qaprotours.service.impl.Scheduler;
 import com.solvd.qaprotours.service.impl.TicketServiceImpl;
 import com.solvd.qaprotours.service.impl.TourServiceImpl;
 import com.solvd.qaprotours.service.impl.fake.FakeAuthService;
+import com.solvd.qaprotours.service.impl.fake.FakeHotelService;
 import com.solvd.qaprotours.service.impl.fake.FakeJwtService;
 import com.solvd.qaprotours.service.impl.fake.FakeTicketService;
 import com.solvd.qaprotours.service.impl.fake.FakeTourService;
@@ -20,10 +24,10 @@ import com.solvd.qaprotours.service.impl.fake.FakeUserClient;
 import com.solvd.qaprotours.service.property.ImageProperties;
 import com.solvd.qaprotours.service.property.JwtProperties;
 import com.solvd.qaprotours.service.property.MinioProperties;
-import com.solvd.qaprotours.web.kafka.MessageSender;
-import com.solvd.qaprotours.web.kafka.MessageSenderImpl;
 import com.solvd.qaprotours.web.mapper.MailDataMapper;
 import com.solvd.qaprotours.web.mapper.MailDataMapperImpl;
+import com.solvd.qaprotours.web.mapper.TourMapper;
+import com.solvd.qaprotours.web.mapper.TourMapperImpl;
 import com.solvd.qaprotours.web.mapper.UserMapper;
 import com.solvd.qaprotours.web.mapper.UserMapperImpl;
 import io.minio.MinioClient;
@@ -92,7 +96,7 @@ public class TestConfig {
 
     @Bean("tourServiceImpl")
     public TourServiceImpl tourServiceImpl() {
-        return new TourServiceImpl(tourRepository);
+        return new TourServiceImpl(tourRepository, messageSender(), tourMapper(), hotelService());
     }
 
     @Primary
@@ -144,6 +148,12 @@ public class TestConfig {
 
     @Primary
     @Bean
+    public HotelService hotelService() {
+        return new FakeHotelService();
+    }
+
+    @Primary
+    @Bean
     public MailDataMapper mailDataMapper() {
         return new MailDataMapperImpl();
     }
@@ -152,6 +162,12 @@ public class TestConfig {
     @Bean
     public UserMapper userMapper() {
         return new UserMapperImpl();
+    }
+
+    @Primary
+    @Bean
+    public TourMapper tourMapper() {
+        return new TourMapperImpl();
     }
 
     @Primary
